consumer/requirements.txt
redis==5.0.1
pika==1.3.2
kafka-python==2.0.2
